+++ src/main/scala/com/easyforger/recipes/CraftingRecipe.scala
@@ -7,18 +7,14 @@ package com.easyforger.recipes
 import net.minecraft.block.Block
 import net.minecraft.item.{Item, ItemStack}
 
-case class CraftingRecipe(sources: Set[RecipeItemStack], shape: Option[String] = None, result: Option[ItemStack] = None) extends RecipeSupport {
+case class CraftingRecipe(sources: Set[RecipeSource], shape: Option[String] = None, result: Option[RecipeResult] = None) extends RecipeOps {
   def +(block: Block): CraftingRecipe = this.copy(sources = sources + block)
   def +(item: Item): CraftingRecipe = this.copy(sources = sources + item)
-  def +(recipeItemStack: RecipeItemStack): CraftingRecipe = this.copy(sources = sources + recipeItemStack)
+  def +(recipeSource: RecipeSource): CraftingRecipe = this.copy(sources = sources + recipeSource)
 
-  def to(item: Item): CraftingRecipe = this.copy(result = Some(new ItemStack(item)))
-  def to(block: Block): CraftingRecipe = this.copy(result = Some(new ItemStack(block)))
-  def to(itemStack: ItemStack): CraftingRecipe = this.copy(result = Some(itemStack))
-
-  // FIXME: this removes an acronym that might have been created, but it is used in a result position, so this is probably not a big deal
-  // still, this kind of thing should not be possible
-  def to(recipeItemStack: RecipeItemStack): CraftingRecipe = this.copy(result = Some(recipeItemStack.itemStack))
+  def to(item: Item): CraftingRecipe = this.copy(result = Some(item))
+  def to(block: Block): CraftingRecipe = this.copy(result = Some(block))
+  def to(itemStack: ItemStack): CraftingRecipe = this.copy(result = Some(RecipeResult(itemStack)))
 
   def withShape(shape: String): CraftingRecipe = this.copy(shape = Some(shape))
 }